# Conflicts:
#	src/Main.java

+++ src/CalculateBookScoreAvg.java
@@ -4,9 +4,9 @@ public class CalculateBookScoreAvg implements CalculateBookScore {
 
     @Override
     public double calculate(List<Book> books) {
-        double sum = 0L;
+        double sum = 0;
         for (Book book : books) {
-            sum += (long) book.getScore();
+            sum += book.getScore();
         }
         return sum/books.size();
     }
